LDrawLoader: Add WebGPU support.
